Fix current cut-out to be motor shield dependent.
The Arduino and Pololu motor shields have very different current sense ratios
(1.65V/A vs 0.525V/A respectively), this means that the Pololu board
effectively gets no current protection from the sketch with the current
settings. That might be OK because the Pololu has built-in short-circuit
protection that generally acts faster than the Arduino's sampling can at
current rates, however I feel that fixing this is the right thing to do anyway.

Also added documentation to describe how the numbers are arrived at.