Support motor shield status feedback.
The pololu motor shield has a feedback pin that reports if the shield has
faulted for any reason (e.g. over current). This is a useful pin to monitor and
plumb to the UI because the Pololu current cut-out is much faster than the
BaseStation, so in the event of a short-circuit it's usually the Pololu that
cuts out first, leaving the tracks without power with no detected current spike
in the UI.

This wires up digital pin 5 to be pulled high, which the board can then pull
low on a fault, which is transmitted to the UI via the new <psf> message.