Avoid transient short-circuits cutting power fully
The pololu internal short-circuit detection is extremely sensitive, so when it
trips we reset the shield once, and only fully disable our power if it trips
again immediately. This is a calculated risk based on how fast the board trips
that re-enabling it once is not enough to cause any damage, and it allows much
smoother running of trains passing over points, etc when the fault is just
transitory.